Registro de Datos Funcional a partir del JSON

+++ src/main/java/com/example/apiregistrousuarios/repository/UsserCRUDRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface UsserCRUDRepository extends CrudRepository<Usser, Long> {
+public interface UsserCRUDRepository extends CrudRepository<Usser, Integer> {
     List<Usser> findByEmail(String email);
     //Optional<User> findByEmail(String email);
